add: cland tidy to cmake config (#189)
* add: clang-tidy to cmake

* fix: repo formating

* fix: bad written exception

* fix: multiple clang-tidy warning

* fix: formating issue

+++ client/src/ClientTickRate.hpp
@@ -7,11 +7,12 @@
 
 #pragma once
 
+#include <cstdint>
 #include <unordered_map>
 
 namespace rtc {
 
-enum class TickRate : std::size_t {
+enum class TickRate : std::uint8_t {
     ENTITY_MOVEMENT,
     CALL_NETWORK_CALLBACKS
 };
